Modularize popularity job
